+++ CHANGES.md
@@ -16,6 +16,8 @@
 - Fixed a bug where the search box was missing for small screen devices.
 - Prevent user adding empty web url
 - Fix bug where search results shown in `My Data` tab
+- Fix bug in function createDiscreteTimesFromIsoSegments where it might create duplicate timestamps.
+- [The next improvement]
 
 #### 8.4.1 - 2023-12-08